Support decimal prices (cents) end-to-end (#21)
Migrate price columns from BIGINT to NUMERIC(20,2), update Go models
from int64 to float64, and make the frontend price input preserve
decimals so pasting "19090.00" from Janice works correctly.

+++ internal/controllers/analytics_test.go
@@ -71,7 +71,7 @@ func TestAnalyticsController_GetSalesMetrics(t *testing.T) {
 		require.NotNil(t, result)
 
 		metrics := result.(*models.SalesMetrics)
-		assert.Equal(t, int64(100000), metrics.TotalRevenue)
+		assert.Equal(t, float64(100000), metrics.TotalRevenue)
 		assert.Equal(t, int64(1), metrics.TotalTransactions)
 		assert.Equal(t, int64(100), metrics.TotalQuantitySold)
 	})
@@ -90,7 +90,7 @@ func TestAnalyticsController_GetSalesMetrics(t *testing.T) {
 		require.NotNil(t, result)
 
 		metrics := result.(*models.SalesMetrics)
-		assert.Equal(t, int64(100000), metrics.TotalRevenue)
+		assert.Equal(t, float64(100000), metrics.TotalRevenue)
 	})
 }
